Improve slide formatting
This commit improves the formatting for the slides, by adding a
background to the title slide and changing the default colors up
a bit. Addtionally there are a couple of other smaller tweaks made.
The theme now matches up pretty closely with the official templates[1]
the only thing missing is the image in the bottom corner, which I
don't view as being critical.

[1] https://www.openstack.org/assets/marketing/presentations/GenericPlatformDeckMay-23-2013.pptx